requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
python-dotenv
SQLAlchemy
pymysql
